+++ src/index.ts
@@ -472,7 +472,7 @@ export class Translator {
         }
         const headers = {
             Authorization: `DeepL-Auth-Key ${authKey}`,
-            'User-Agent': 'deepl-node/1.6.0',
+            'User-Agent': 'deepl-node/1.7.0',
             ...(options?.headers ?? {}),
         };