+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [1.7.0] - 2022-09-30
 ### Added
 * Add formality options `'prefer_less'` and `'prefer_more'`.
 ### Changed
@@ -122,7 +122,7 @@ client library took over this package name. Thanks to
 ownership.
 
 
-[Unreleased]: https://github.com/DeepLcom/deepl-node/compare/v1.6.0...HEAD
+[1.7.0]: https://github.com/DeepLcom/deepl-node/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/DeepLcom/deepl-node/compare/v1.5.0...v1.6.0
 [1.5.0]: https://github.com/DeepLcom/deepl-node/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/DeepLcom/deepl-node/compare/v1.3.2...v1.4.0